Modularization of task scheduling
